+++ weblate/trans/views/create.py
@@ -125,7 +125,8 @@ class CreateProject(BaseCreateView):
 
     def can_create(self):
         return (
-            (self.has_billing and self.billings) or self.request.user.is_superuser
+            (self.has_billing and self.billings)
+            or self.request.user.has_perm('project.add')
         )
 
     def post(self, request, *args, **kwargs):